--- src/dictionary/unit-1/audio1_1.ts
@@ -1,7 +1,0 @@
-import c_1_2 from "../../assets/sound-conversation/c_1_2.mp3"
-export const audio1_1 = [
-  { 'id':1,
-    'text':'-Chào chị! Tên chị là gì?\n-Tên tôi là Mary.\n-Chỉ là người nước nào?\n-Tôi là người Mỹ, còn em?\n-Em là ngã người Việt Nam.\n-Rất vui được gặp em!\n-Rất vui được gặp chị!',
-    'audio': c_1_2,
-  }
-];
