--- src/dictionary/unit-2/audio2_1.ts
@@ -1,11 +1,0 @@
-
-import c_2_2 from "../../assets/sound-conversation/c_2_2.mp3"
-
-export const audio2_1 = [
-
-  {
-    'text':'-Chào anh Frank, đây là Son, em trai tôi.\n-Chào em Lan và em Sơn. Đây là ông George, thầy giáo của tôi.\n-Chào anh Frank và thấy George. rat vui duoc gap thay va anh.\n-Cám ơn em! Thưa thầy George! Kia là Lan, cô giáo tiếng Việt của tôi.\n-Xin chào! Tôi là George đến từ Anh. Rất vui được gặp các em.',
-    'audio':c_2_2,
-  },
-
-];
