--- src/dictionary/unit-2/audio2_2.ts
@@ -1,9 +1,0 @@
-
-import c_2_3 from "../../assets/sound-conversation/c_2_3.mp3"
-export const audio2_2 = [
-
-  {
-    'text':'-Chào chị, lâu rồi không gặp chị! Chị có khỏe không?\n-Chị khoe, cảm ơn Lần. Thế còn em?\n-Dạo này, em không được khỏe lắm.\n-Thế à! Em giữ sức khỏe nhé!\n-Vàng! cảm ơn chị ạ!',
-    'audio':c_2_3,
-  }
-];
